+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Project uses [Semantic Versioning](http://semver.org/)
 ###Added
 - UTC Clock to the schedule
 
+###Fixed
+- "Clear hidden" button now only renders on the General tab of options
+
 ##0.11.0 - 2015-15-03
 ###Added
 - PDF uploads