+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Project uses [Semantic Versioning](http://semver.org/)
 ###Added
 - UTC Clock to the schedule
 - Options export and import to/from file
+- Option to anonymise all posters
 
 ###Fixed
 - "Clear hidden" button now only renders on the General tab of options